refactor code take app out of server
